Fix upload mistake
I erroneously changed the original file and not the one in my git repository, so I uploaded the original with no changes. Upload the one which was originally meant for upload and put it in a folder so Arduino IDE shuts up about it not being in one and actually open it (which caused the mistake in the first place).